requirements.txt
streamlit
qdrant-client
sentence-transformers
groq
pandas
python-dotenv